+++ aksel.nav.no/website/app/api/markdown/blocks/KodeEksempler.md.tsx
@@ -1,5 +1,5 @@
 import { type ExtractPortableMarkdownComponentProps } from "@/app/_sanity/types";
-import { buildMarkdown } from "@/app/api/markdown/utils";
+import { buildMarkdown } from "@/app/api/markdown/helpers/build-markdown";
 
 /**
  * TODO:
